+++ src/ui/widgets/_init.lua
@@ -1,5 +1,5 @@
 local _, Addon = ...
-local Widgets = Addon:GetModule("Widgets")
+local Widgets = Addon:GetModule("Widgets") ---@class Widgets
 
 Widgets.BORDER_BACKDROP = {
   bgFile = "Interface\\Buttons\\WHITE8x8",
@@ -13,6 +13,9 @@ do -- Widget:Padding()
   local PADDING = 8
   local cache = {}
 
+  --- Returns a base padding of `8` multiplied by the given number.
+  --- @param multiplier? number
+  --- @return number padding
   function Widgets:Padding(multiplier)
     if type(multiplier) ~= "number" then return PADDING end
     local value = cache[tostring(multiplier)]