+++ src/ui/widgets/frame.lua
@@ -1,20 +1,27 @@
 local _, Addon = ...
-local Tooltip = Addon:GetModule("Tooltip")
-local Widgets = Addon:GetModule("Widgets")
+local Tooltip = Addon:GetModule("Tooltip") ---@type Tooltip
+local Widgets = Addon:GetModule("Widgets") ---@class Widgets
 
---[[
-  Creates a basic frame with a backdrop.
+-- =============================================================================
+-- LuaCATS Annotations
+-- =============================================================================
 
-  options = {
-    name? = string,
-    frameType? = string,
-    parent? = UIObject,
-    points? = table[],
-    width? = number,
-    height? = number,
-    onUpdateTooltip? = function(self, tooltip) -> nil
-  }
-]]
+--- @class FrameWidgetOptions
+--- @field name? string
+--- @field frameType? string
+--- @field parent? table
+--- @field points? table[]
+--- @field width? integer
+--- @field height? integer
+--- @field onUpdateTooltip? fun(self: FrameWidget, tooltip: Tooltip)
+
+-- =============================================================================
+-- Widgets - Frame
+-- =============================================================================
+
+--- Creates a basic frame with a backdrop.
+--- @param options FrameWidgetOptions
+--- @return FrameWidget frame
 function Widgets:Frame(options)
   -- Defaults.
   options.frameType = Addon:IfNil(options.frameType, "Frame")
@@ -23,13 +30,13 @@ function Widgets:Frame(options)
   options.height = Addon:IfNil(options.height, 1)
 
   -- Base frame.
-  local frame = CreateFrame(options.frameType, options.name, options.parent)
+  local frame = CreateFrame(options.frameType, options.name, options.parent) ---@class FrameWidget
   frame:SetClipsChildren(true)
 
   -- Backdrop.
   Mixin(frame, BackdropTemplateMixin)
   frame:SetBackdrop(self.BORDER_BACKDROP)
-  frame:SetBackdropColor(0, 0, 0, 0.75)
+  frame:SetBackdropColor(0.05, 0.05, 0.05, 0.95)
   frame:SetBackdropBorderColor(0, 0, 0, 1)
 
   -- Size.