+++ src/ui/widgets/text-frame.lua
@@ -1,6 +1,6 @@
 local _, Addon = ...
-local L = Addon:GetModule("Locale")
-local Widgets = Addon:GetModule("Widgets")
+local L = Addon:GetModule("Locale") ---@type Locale
+local Widgets = Addon:GetModule("Widgets") ---@class Widgets
 
 --[[
   Creates a ScrollableTitleFrame with an edit box.